# src/config/apispec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Events api",
    "description": "API Docs for Events management app built with express js",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://events-be.onrender.com/api/v1",
      "description": "Live server"
    },
    {
      "url": "http://localhost:7000/api/v1",
      "description": "Dev server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/user/signup": {
      "post": {
        "tags": ["user"],
        "summary": "signup user",
        "description": "Endpoint to add new users",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/user/signin": {
      "post": {
        "tags": ["user"],
        "summary": "signin user",
        "description": "Endpoint to sign in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "example": {
                "email": "admin@mail.com",
                "password": "admin123"
              }
            }
          }
        },
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/events/": {
      "get": {
        "tags": ["event"],
        "summary": "Retrieve all events",
        "description": "Endpoint to retrieve all events",
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "post": {
        "tags": ["event"],
        "summary": "create event",
        "description": "Endpoint to add new event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              },
              "example": {
                "name": "updated stuff 3",
                "description": "here is a little event info",
                "start_date": "2023-04-30T02:15:12.356Z",
                "end_date": "2023-05-30T02:15:12.356Z",
                "isPremium": false,
                "location": "online",
                "speakers": [
                  {
                    "name": "Judge Bush",
                    "designation": "President"
                  },
                  {
                    "name": "Jane Lane",
                    "designation": "VP"
                  }
                ],
                "event_types": ["stuff", "new"]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "default": {
            "description": "sucessful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "delete": {
        "tags": ["event"],
        "summary": "delete event",
        "description": "Endpoint to delete event by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["event"],
        "summary": "event by ID",
        "description": "Endpoint to rertieve event by ID",
        "security": [
          {
            "noauthAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["event"],
        "summary": "update event by ID",
        "description": "Endpoint toupdate event by ID",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event",
                "example": {
                  "name": "updated stuff 3",
                  "start_date": "2023-04-30T02:15:12.356Z",
                  "end_date": "2023-05-30T02:15:12.356Z",
                  "isPremium": false,
                  "speakers": [
                    {
                      "name": "Judge Bush",
                      "designation": "President"
                    },
                    {
                      "name": "Jane Lane",
                      "designation": "VP"
                    }
                  ],
                  "event_types": ["stuff", "new"]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": "{\r\n    \"status\": \"successful\",\r\n    \"response\": {\r\n        \"id\": \"4986b3a3-5ddd-4ac2-8b27-c24af2cc0ae9\",\r\n        \"name\": \"updated\",\r\n        \"description\": \"Event like no other\",\r\n        \"start_date\": \"2023-04-30T02:15:12.356Z\",\r\n        \"end_date\": \"2023-05-30T02:15:12.356Z\",\r\n        \"isPremium\": true,\r\n        \"author_id\": \"ab845850-925e-4138-91b1-f6d54d377cbb\",\r\n        \"createdAt\": \"2022-12-31T02:32:30.617Z\",\r\n        \"updatedAT\": \"2023-01-13T21:07:17.244Z\"\r\n    }\r\n}"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "noauthAuth": {
        "type": "http",
        "scheme": "noauth"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "id": {
            "type": "string",
            "description": "auto generated uuid"
          },
          "name": {
            "type": "string",
            "description": "user name"
          },
          "email": {
            "type": "string",
            "description": "user email address"
          },
          "password": {
            "type": "string",
            "description": "user password"
          }
        },
        "example": {
          "name": "John Doe",
          "email": "john@mail.com",
          "password": "pass123"
        }
      },
      "Event": {
        "type": "object",
        "required": ["name", "start_date", "end_date"],
        "properties": {
          "id": {
            "type": "string",
            "description": "auto generated uuid"
          },
          "name": {
            "type": "string",
            "description": "Event name"
          },
          "description": {
            "type": "string",
            "description": "Event description"
          },
          "start_date": {
            "type": "string",
            "description": "Event start date and time"
          },
          "end_date": {
            "type": "string",
            "description": "Event end date and time"
          },
          "isPremium": {
            "type": "boolean",
            "description": "Confirm if event is premium"
          },
          "event_type": {
            "type": "object",
            "description": "Array of object"
          },
          "speakers": {
            "type": "object",
            "description": "Array of object"
          },
          "author_id": {
            "type": "string",
            "description": "auto generated uuid"
          },
          "createdAt": {
            "type": "string",
            "description": "auto generated date"
          },
          "updatedAt": {
            "type": "string",
            "description": "auto generated date"
          }
        },
        "example": {
          "name": "Event of the year",
          "description": "This is an event of a life time, you don't want to miss It",
          "start_date": "2023-04-30T02:15:12.356Z",
          "end_date": "2023-04-30T09:15:12.356Z",
          "isPremium": true,
          "event_types": ["fun", "hangout"],
          "speakers": [
            {
              "name": "John Doe",
              "designation": "Host"
            }
          ]
        }
      }
    }
  }
}